pyproject.toml
[project]
name = "ui-test-20260302-v3"
version = "0.1.0"
description = "Chain-of-Thought prompt tuning experiments"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "wandb>=0.16.0",
    "openai>=1.12.0",
    "datasets>=2.16.0",
    "numpy>=1.24.0",
    "matplotlib>=3.8.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"